requirements.txt
ariestools==1.4.0
PyInstaller==3.6
psycopg2-binary==2.8.5